requirements.txt
asgiref==3.8.1
bleach==6.1.0
cloudinary==1.41.0
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
dj3-cloudinary-storage==0.0.6
Django==4.2.9
django-allauth==65.0.1
django-axes==6.5.2
django-crispy-forms==2.3
django-summernote==0.8.20.0
gunicorn==23.0.0
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests-oauthlib==2.0.0
sqlparse==0.5.1
webencodings==0.5.1
whitenoise==6.7.0
